third_party/WebKit/LayoutTests/fast/dom/Element/setAttributeNode-for-existing-attribute-expected.txt

Testcase for bug 279193: setAttributeNode() does not set the new value to an existing attribute if specified attribute is in a different case. The issue can be manually verified by ascertaining that a green colored block is displayed instead of a red one.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testElement.attributes.length is 2
PASS testElement.style.backgroundColor is "green"
PASS successfullyParsed is true

TEST COMPLETE
